[FIX] Missing error handling, CT counting bug
 - The new FASTQ I/O didn't have error handling
 - The new workflow didn't take into account the size of multi records
   for building the clonestore, i.e. counted every result only once

+++ src/fastq_multi_record_types.h
@@ -54,7 +54,7 @@ struct FastqMultiRecord {};
 template<>
 struct FastqMultiRecord<SingleEnd> {
     typedef String<Dna5> TSequence;
-    typedef String<float> TQualities;
+    typedef String<double> TQualities;
     typedef std::set<CharString> TIds;
 
     TSequence   seq, bcSeq;
@@ -68,7 +68,7 @@ struct FastqMultiRecord<SingleEnd> {
 template<>
 struct FastqMultiRecord<PairedEnd> {
     typedef String<Dna5> TSequence;
-    typedef String<float> TQualities;
+    typedef String<double> TQualities;
     typedef std::set<CharString> TIds;
 
     TSequence   fwSeq, revSeq, bcSeq;
